Move so3_surface to solid_3D_ele
