+++ src/cardiovascular0d/4C_cardiovascular0d.cpp
@@ -16,7 +16,6 @@
 #include "4C_linalg_serialdensevector.hpp"
 #include "4C_linalg_utils_densematrix_multiply.hpp"
 #include "4C_linalg_utils_sparse_algebra_math.hpp"
-#include "4C_so3_surface.hpp"
 
 #include <Teuchos_StandardParameterEntryValidators.hpp>
