+++ src/cardiovascular0d/4C_cardiovascular0d_arterialproxdist.cpp
@@ -16,7 +16,6 @@
 #include "4C_linalg_serialdensematrix.hpp"
 #include "4C_linalg_serialdensevector.hpp"
 #include "4C_linalg_utils_sparse_algebra_assemble.hpp"
-#include "4C_so3_surface.hpp"
 #include "4C_utils_function_of_time.hpp"
 
 #include <iostream>
