+++ src/cardiovascular0d/4C_cardiovascular0d_respiratory_syspulperiphcirculation.cpp
@@ -12,7 +12,6 @@
 #include "4C_linalg_serialdensematrix.hpp"
 #include "4C_linalg_serialdensevector.hpp"
 #include "4C_linalg_utils_sparse_algebra_assemble.hpp"
-#include "4C_so3_surface.hpp"
 #include "4C_utils_function_of_time.hpp"
 
 #include <iostream>
