+++ src/solid_3D_ele/4C_solid_3D_ele.cpp
@@ -15,12 +15,12 @@
 #include "4C_io_input_spec.hpp"
 #include "4C_io_input_spec_builders.hpp"
 #include "4C_mat_so3_material.hpp"
-#include "4C_so3_surface.hpp"
 #include "4C_solid_3D_ele_factory.hpp"
 #include "4C_solid_3D_ele_interface_serializable.hpp"
 #include "4C_solid_3D_ele_line.hpp"
 #include "4C_solid_3D_ele_nullspace.hpp"
 #include "4C_solid_3D_ele_properties.hpp"
+#include "4C_solid_3D_ele_surface.hpp"
 #include "4C_solid_3D_ele_utils.hpp"
 #include "4C_structure_new_elements_paramsinterface.hpp"
 
@@ -216,7 +216,7 @@ std::vector<std::shared_ptr<Core::Elements::Element>> Discret::Elements::Solid::
 
 std::vector<std::shared_ptr<Core::Elements::Element>> Discret::Elements::Solid::surfaces()
 {
-  return Core::Communication::get_element_surfaces<StructuralSurface, Solid>(*this);
+  return Core::Communication::get_element_surfaces<SolidSurface, Solid>(*this);
 }
 
 const Core::FE::GaussIntegration& Discret::Elements::Solid::get_gauss_rule() const
