+++ src/solid_poro_3D_ele/4C_solid_poro_3D_ele_pressure_velocity_based.cpp
@@ -16,11 +16,11 @@
 #include "4C_io_input_spec_builders.hpp"
 #include "4C_mat_fluidporo.hpp"
 #include "4C_mat_structporo.hpp"
-#include "4C_so3_surface.hpp"
 #include "4C_solid_3D_ele_factory.hpp"
 #include "4C_solid_3D_ele_interface_serializable.hpp"
 #include "4C_solid_3D_ele_line.hpp"
 #include "4C_solid_3D_ele_nullspace.hpp"
+#include "4C_solid_3D_ele_surface.hpp"
 #include "4C_solid_3D_ele_utils.hpp"
 #include "4C_solid_poro_3D_ele_factory.hpp"
 #include "4C_solid_poro_3D_ele_utils.hpp"
@@ -183,8 +183,8 @@ Discret::Elements::SolidPoroPressureVelocityBased::lines()
 std::vector<std::shared_ptr<Core::Elements::Element>>
 Discret::Elements::SolidPoroPressureVelocityBased::surfaces()
 {
-  return Core::Communication::get_element_surfaces<StructuralSurface,
-      SolidPoroPressureVelocityBased>(*this);
+  return Core::Communication::get_element_surfaces<SolidSurface, SolidPoroPressureVelocityBased>(
+      *this);
 }
 
 void Discret::Elements::SolidPoroPressureVelocityBased::set_params_interface_ptr(
