+++ src/solid_poro_3D_ele/4C_solid_poro_3D_ele_pressure_velocity_based_p1.cpp
@@ -17,10 +17,10 @@
 #include "4C_io_input_spec_builders.hpp"
 #include "4C_mat_fluidporo.hpp"
 #include "4C_mat_structporo.hpp"
-#include "4C_so3_surface.hpp"
 #include "4C_solid_3D_ele_interface_serializable.hpp"
 #include "4C_solid_3D_ele_line.hpp"
 #include "4C_solid_3D_ele_properties.hpp"
+#include "4C_solid_3D_ele_surface.hpp"
 #include "4C_solid_3D_ele_utils.hpp"
 #include "4C_solid_poro_3D_ele_factory.hpp"
 
@@ -181,8 +181,8 @@ Discret::Elements::SolidPoroPressureVelocityBasedP1::lines()
 std::vector<std::shared_ptr<Core::Elements::Element>>
 Discret::Elements::SolidPoroPressureVelocityBasedP1::surfaces()
 {
-  return Core::Communication::get_element_surfaces<StructuralSurface,
-      SolidPoroPressureVelocityBasedP1>(*this);
+  return Core::Communication::get_element_surfaces<SolidSurface, SolidPoroPressureVelocityBasedP1>(
+      *this);
 }
 
 void Discret::Elements::SolidPoroPressureVelocityBasedP1::set_params_interface_ptr(
